[move/#71] domain으로 모델 위치 변경
