+++ domain/gallery/src/main/java/com/together/study/gallery/model/GalleryImage.kt
@@ -1,0 +1,6 @@
+package com.together.study.gallery.model
+
+data class GalleryImage(
+    val id: Long,
+    val dateMillis: Long,
+)
