+++ presentation/gallery/build.gradle.kts
@@ -15,6 +15,8 @@ dependencies {
     implementation(projects.core.designsystem)
     implementation(projects.core.util)
 
+    implementation(projects.domain.gallery)
+
     implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.navigation.compose)
     implementation(libs.coil.compose)
